models: move mimetypes to models
