+++ models/mime.go
@@ -1,9 +1,12 @@
-package lens
+package models
 
 // MimeType are categories recognized by Lens
 type MimeType string
 
 const (
+	// MimeTypeUnknown is a catch-all for unspecified object type
+	MimeTypeUnknown = "unknown"
+
 	// MimeTypePDF is a pdf document
 	MimeTypePDF = "pdf"
 	// MimeTypeDocument is a plain-text document
